pyproject.toml
[tool.poetry]
name = "amazing-hunting"
version = "0.1.0"
description = ""
authors = ["zacadaaa <styopanikonov@yandex.ru>"]
readme = "README.md"
packages = [{include = "amazing_hunting"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
psycopg2 = "^2.9.5"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
